requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
scikit-learn
httpx
groq
pymupdf
streamlit
pandas
